Move Layout.tsx to 3d panel top level folder
