+++ packages/studio-base/src/panels/ThreeDimensionalViz/Layout.tsx
