+++ packages/studio-base/src/panels/ThreeDimensionalViz/LayoutToolbar.tsx
@@ -20,6 +20,7 @@ import Crosshair from "@foxglove/studio-base/panels/ThreeDimensionalViz/Crosshai
 import FollowTFControl from "@foxglove/studio-base/panels/ThreeDimensionalViz/FollowTFControl";
 import Interactions from "@foxglove/studio-base/panels/ThreeDimensionalViz/Interactions";
 import { TabType } from "@foxglove/studio-base/panels/ThreeDimensionalViz/Interactions/Interactions";
+import { LayoutToolbarSharedProps } from "@foxglove/studio-base/panels/ThreeDimensionalViz/Layout";
 import MainToolbar from "@foxglove/studio-base/panels/ThreeDimensionalViz/MainToolbar";
 import MeasureMarker from "@foxglove/studio-base/panels/ThreeDimensionalViz/MeasureMarker";
 import MeasuringTool, {
@@ -28,7 +29,6 @@ import MeasuringTool, {
 import SearchText, {
   SearchTextProps,
 } from "@foxglove/studio-base/panels/ThreeDimensionalViz/SearchText";
-import { LayoutToolbarSharedProps } from "@foxglove/studio-base/panels/ThreeDimensionalViz/TopicTree/Layout";
 import { fonts } from "@foxglove/studio-base/util/sharedStyleConstants";
 
 type Props = LayoutToolbarSharedProps &
